@@ -67,10 +67,10 @@
     "lodash": "^4.3.0",
     "mocha": "^3.0.2",
     "node-sass": "^3.4.2",
-    "react": "^15.0.0",
+    "react": "^16.0.0",
     "react-addons-test-utils": "^15.0.0",
     "react-docgen": "^2.7.0",
-    "react-dom": "^15.0.0",
+    "react-dom": "^16.0.0",
     "react-transform-hmr": "^1.0.4",
     "sass-loader": "^4.0.0",
     "sinon": "^1.15.4",
@@ -81,11 +81,12 @@
     "webpack-hot-middleware": "^2.10.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "dependencies": {
     "classnames": "^2.2.1",
+    "create-react-class": "^15.6.3",
     "moment": "^2.13.0",
     "react-onclickoutside": "^4.8.0",
     "tether": "^1.3.2"
