@@ -3,10 +3,10 @@ import React from 'react'
 import classnames from 'classnames'
 import { isSameDay, isDayDisabled, isDayInRange } from './date_utils'
 
-var Day = React.createClass({
-  displayName: 'Day',
+class Day extends React.Component {
+  static displayName = 'Day';
 
-  propTypes: {
+  static propTypes = {
     day: React.PropTypes.object.isRequired,
     endDate: React.PropTypes.object,
     excludeDates: React.PropTypes.array,
@@ -24,45 +24,45 @@ var Day = React.createClass({
     selectsStart: React.PropTypes.bool,
     startDate: React.PropTypes.object,
     utcOffset: React.PropTypes.number
-  },
-  getDefaultProps () {
-    return {
-      utcOffset: moment.utc().utcOffset()
-    }
-  },
-  handleClick (event) {
+  };
+
+  static defaultProps = {
+    utcOffset: moment.utc().utcOffset()
+  };
+
+  handleClick = (event) => {
     if (!this.isDisabled() && this.props.onClick) {
       this.props.onClick(event)
     }
-  },
+  };
 
-  handleMouseEnter (event) {
+  handleMouseEnter = (event) => {
     if (!this.isDisabled() && this.props.onMouseEnter) {
       this.props.onMouseEnter(event)
     }
-  },
+  };
 
-  isSameDay (other) {
+  isSameDay = (other) => {
     return isSameDay(this.props.day, other)
-  },
+  };
 
-  isDisabled () {
+  isDisabled = () => {
     return isDayDisabled(this.props.day, this.props)
-  },
+  };
 
-  isHighlighted () {
+  isHighlighted = () => {
     const { day, highlightDates } = this.props
     if (!highlightDates) return false
     return highlightDates.some((testDay) => { return isSameDay(day, testDay) })
-  },
+  };
 
-  isInRange () {
+  isInRange = () => {
     const { day, startDate, endDate } = this.props
     if (!startDate || !endDate) return false
     return isDayInRange(day, startDate, endDate)
-  },
+  };
 
-  isInSelectingRange () {
+  isInSelectingRange = () => {
     const { day, selectsStart, selectsEnd, selectingDate, startDate, endDate } = this.props
 
     if (!(selectsStart || selectsEnd) || !selectingDate || this.isDisabled()) {
@@ -78,9 +78,9 @@ var Day = React.createClass({
     }
 
     return false
-  },
+  };
 
-  isSelectingRangeStart () {
+  isSelectingRangeStart = () => {
     if (!this.isInSelectingRange()) {
       return false
     }
@@ -92,9 +92,9 @@ var Day = React.createClass({
     } else {
       return isSameDay(day, startDate)
     }
-  },
+  };
 
-  isSelectingRangeEnd () {
+  isSelectingRangeEnd = () => {
     if (!this.isInSelectingRange()) {
       return false
     }
@@ -106,31 +106,31 @@ var Day = React.createClass({
     } else {
       return isSameDay(day, endDate)
     }
-  },
+  };
 
-  isRangeStart () {
+  isRangeStart = () => {
     const { day, startDate, endDate } = this.props
     if (!startDate || !endDate) return false
     return isSameDay(startDate, day)
-  },
+  };
 
-  isRangeEnd () {
+  isRangeEnd = () => {
     const { day, startDate, endDate } = this.props
     if (!startDate || !endDate) return false
     return isSameDay(endDate, day)
-  },
+  };
 
-  isWeekend () {
+  isWeekend = () => {
     const weekday = this.props.day.day()
     return weekday === 0 || weekday === 6
-  },
+  };
 
-  isOutsideMonth () {
+  isOutsideMonth = () => {
     return this.props.month !== undefined &&
       this.props.month !== this.props.day.month()
-  },
+  };
 
-  getClassNames () {
+  getClassNames = () => {
     return classnames('react-datepicker__day', {
       'react-datepicker__day--disabled': this.isDisabled(),
       'react-datepicker__day--selected': this.isSameDay(this.props.selected),
@@ -145,9 +145,9 @@ var Day = React.createClass({
       'react-datepicker__day--weekend': this.isWeekend(),
       'react-datepicker__day--outside-month': this.isOutsideMonth()
     })
-  },
+  };
 
-  render () {
+  render() {
     return (
       <div
           className={this.getClassNames()}
@@ -159,6 +159,6 @@ var Day = React.createClass({
       </div>
     )
   }
-})
+}
 
 module.exports = Day
